\name{openOutputFileHandle}
\alias{openOutputFileHandle}
\title{openOutputFileHandle}
\usage{
  openOutputFileHandle(input_file, suffix, header,
    overwrite)
}
\arguments{
  \item{input_file}{}

  \item{suffix}{}

  \item{header}{}

  \item{overwrite}{}
}
\value{
  open file handle
}
\description{
  Open file handle for writing output.
}
\author{
  Adam Struck - Intern
}

