\name{writeResults}
\alias{writeResults}
\title{writeResults}
\usage{
  writeResults(results.table, output.fh)
}
\arguments{
  \item{results.table}{}

  \item{output.fh}{}
}
\value{
  tsv file
}
\description{
  Write a data frame to a file handle
}
\details{
  .. content for \details{} ..
}
\author{
  Adam Struck - Intern
}

